Add files via upload
this is the end process of how I discovered the UPG Universal Pattern Generator equation P(T)= 2 + 2e^(iMT)(cos(m)·MT + i sin(MT)), this equation is the foundation of all interference patterns,  it encodes all quantum mechanics and is itself the Light Cone,  it shows the emergence of Numbers including Primes, how they are created, through the Cardioid.  T=[0...N] and explains how the roots of unity are formed through the reduction of N to its primary energy level the Fourier kernel and much more.  this is the foundation of the creation of Numbers through the constant Phi.